refactor/fix: use size_t for dynarray, fix a gcc warning

+++ src/dynarray.h
@@ -3,15 +3,15 @@
 
 #define DynArray(T) struct { \
     T *data; \
-    int count; \
-    int capacity; \
+    size_t count; \
+    size_t capacity; \
 }
 
 #define dynarray_insert(array, stmt) \
 do { \
     if ((array)->count >= (array)->capacity) { \
         (array)->capacity = (array)->capacity == 0 ? 2 : (array)->capacity * 2; \
-        (array)->data = realloc((array)->data, sizeof((array)->data[0]) * (array)->capacity); \  
+        (array)->data = realloc((array)->data, sizeof((array)->data[0]) * (array)->capacity); \
     } \
     (array)->data[(array)->count++] = (stmt); \
 } while(0)
